fix wifi station mode
