+++ sdkconfig
@@ -492,13 +492,13 @@ CONFIG_EXAMPLE_WIFI_SCAN_METHOD_ALL_CHANNEL=y
 # WiFi Scan threshold
 #
 CONFIG_EXAMPLE_WIFI_SCAN_RSSI_THRESHOLD=-127
-CONFIG_EXAMPLE_WIFI_AUTH_OPEN=y
+# CONFIG_EXAMPLE_WIFI_AUTH_OPEN is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WEP is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WPA_PSK is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WPA2_PSK is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WPA_WPA2_PSK is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WPA2_ENTERPRISE is not set
-# CONFIG_EXAMPLE_WIFI_AUTH_WPA3_PSK is not set
+CONFIG_EXAMPLE_WIFI_AUTH_WPA3_PSK=y
 # CONFIG_EXAMPLE_WIFI_AUTH_WPA2_WPA3_PSK is not set
 # CONFIG_EXAMPLE_WIFI_AUTH_WAPI_PSK is not set
 # end of WiFi Scan threshold
@@ -754,8 +754,8 @@ CONFIG_ESP_HTTP_CLIENT_ENABLE_HTTPS=y
 #
 # HTTP Server
 #
-CONFIG_HTTPD_MAX_REQ_HDR_LEN=512
-CONFIG_HTTPD_MAX_URI_LEN=512
+CONFIG_HTTPD_MAX_REQ_HDR_LEN=1024
+CONFIG_HTTPD_MAX_URI_LEN=1024
 CONFIG_HTTPD_ERR_RESP_NO_DELAY=y
 CONFIG_HTTPD_PURGE_BUF_LEN=32
 # CONFIG_HTTPD_LOG_PURGE_DATA is not set
