+++ simple_server_Spotify.code-workspace
@@ -1,0 +1,21 @@
+{
+	"folders": [
+		{
+			"path": "."
+		},
+		{
+			"path": "../station"
+		}
+	],
+	"settings": {
+		"files.associations": {
+			"mdns.h": "c",
+			"typeinfo": "c",
+			"main.h": "c",
+			"wificonfig.h": "c",
+			"sockets.h": "c",
+			"esp_wifi.h": "c",
+			"freertos.h": "c"
+		}
+	}
+}
